use_namespaces is now default, still needs the upgrade code

+++ t/backcompat/0.04006/10sqlite_common.t
@@ -21,5 +21,5 @@ my $class = $@ ? 'SQLite2' : 'SQLite';
 }
 
 END {
-    unlink './t/sqlite_test';
+    unlink './t/sqlite_test' if $ENV{SCHEMA_LOADER_TESTS_BACKCOMPAT};
 }
